Move `CourseIdStub` to test namespace
