+++ tests/Context/Video/Module/Video/Domain/VideoStub.php
@@ -10,18 +10,24 @@ use CodelyTv\Context\Video\Module\Video\Domain\VideoTitle;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoType;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoUrl;
 use CodelyTv\Shared\Domain\CourseId;
-use CodelyTv\Shared\Test\Stub\CourseIdStub;
+use CodelyTv\Test\Context\Course\Module\Course\Domain\CourseIdStub;
 
 final class VideoStub
 {
+    public static function withId(VideoId $id)
+    {
+        return self::create(
+            $id,
+            VideoTypeStub::random(),
+            VideoTitleStub::random(),
+            VideoUrlStub::random(),
+            CourseIdStub::random()
+        );
+    }
+
     public static function create(VideoId $id, VideoType $type, VideoTitle $title, VideoUrl $url, CourseId $courseId)
     {
         return new Video($id, $type, $title, $url, $courseId);
-    }
-
-    public static function withId(VideoId $id)
-    {
-        return self::create($id, VideoTypeStub::random(), VideoTitleStub::random(), VideoUrlStub::random(), CourseIdStub::random());
     }
 
     public static function random(): Video
